Fix bug in how the containing polygon is found
This PR adds a new test for a failing pathfind that I discovered while
building a game [1]. `NavMesh.findPath` was returning `null` despite a
trivial solution.

This PR then fixes that bug. I found that `_findContainingPolygon` was
not finding the polygons containing the points, and so concluded there
was no path. The old code looks strange to the eye. After looking at
how the polygon bounds are calculated, I think the old code was just
wrong.

The rest of the testsuite still passes with this fix in place.

[1] https://github.com/46bit/rtt